notifications: cleanup queues periodically
In previous implementation, resources were leaked in case of disconnection
of all clients (sema:count() == 0) and no notification on the channel.
The entry of "queues" in nginx and the channel listener in postres were
leaked.

Function cleanup_queues was introduced. It checks all queues and removes
unused items and unlistens corresponding channels. It is called when
reader light thread detects read timeout (after 5 minutes of no
notifications) or once in 10000 received notifications.

Relistening channels after restart of the background Lua handler
was moved to function listen_queues.